div: enable input buffer to allow more inflights
